[AOS] feat : 글림 조회, 단건 조회

+++ Glim_AOS/presentation/src/main/java/com/jeongbj/presentation/feature/book/detail/BookDetailNavigation.kt
@@ -12,7 +12,7 @@ data class BookDetailRoute(
 )
 
 fun NavGraphBuilder.bookDetailNav(
-    navigateToQuoteDetail: () -> Unit,
+    navigateToQuoteDetail: (Long) -> Unit,
     navigateBack: () -> Unit
 ) {
     composable<BookDetailRoute> {
